Implemented vision api
it only supports web links, for now. In the future it'd be cool to allow for image uploads, pasting and drag-and-dropping, and displaying the images in the chat. A bandaid for that might be to have a way to upload a pasted image to imgur and have the link appear in the chat.